extract GoogleEventExportHelper.
