+++ spec/helpers/events_helper_spec.rb
@@ -44,64 +44,6 @@ describe EventsHelper, :type => :helper do
     end
   end
 
-  describe "google_event_export_link" do
-    def escape(string)
-      return Regexp.escape(CGI.escape(string))
-    end
-
-    shared_context "exported event setup" do
-      before do
-        @venue = Venue.create!(:title => "My venue", :address => "1930 SW 4th Ave, Portland, Oregon 97201")
-        @event = Event.create!(:title => "My event", :start_time => Time.now - 1.hour, :end_time => Time.now, :venue => @venue, :description => event_description)
-        @export = helper.google_event_export_link(@event)
-      end
-    end
-
-    shared_examples_for "exported event" do
-      it "should have title" do
-        expect(@export).to match /\&text=#{escape(@event.title)}/
-      end
-
-      it "should have time range" do
-        expect(@export).to match /\&dates=#{escape(helper.format_google_timespan(@event))}/
-      end
-
-      it "should have venue title" do
-        expect(@export).to match /\&location=#{escape(@event.venue.title)}/
-      end
-
-      it "should have venue address" do
-        expect(@export).to match /\&location=.+?#{escape(@event.venue.geocode_address)}/
-      end
-    end
-
-    describe "an event's text doesn't need truncation" do
-      let(:event_description) { "My event description." }
-      include_context "exported event setup"
-
-      it_should_behave_like "exported event"
-
-      it "should have a complete event description" do
-        expect(@export).to match /\&details=.*#{escape(event_description)}/
-      end
-    end
-
-    describe "an event's text needs truncation" do
-      let(:event_description) { "My event description. " * 100 }
-      include_context "exported event setup"
-
-      it_should_behave_like "exported event"
-
-      it "should have a truncated event description" do
-        expect(@export).to match /\&details=.*#{escape(event_description[0..100])}/
-      end
-
-      it "should have a truncated URL" do
-        expect(@export.size).to be < event_description.size
-      end
-    end
-  end
-
   describe "#google_events_subscription_link" do
     def method(*args)
       helper.google_events_subscription_link(*args)
@@ -198,24 +140,6 @@ describe EventsHelper, :type => :helper do
         end_time: "2010-01-02 12:00:00")
       event.venue = FactoryGirl.create(:venue, title: "holocene")
       expect(tweet_text(event)).to eq("hip and/or hop - 12:00PM 01.01.2010 @ holocene")
-    end
-  end
-
-  describe "format_google_timespan" do
-    it "should use the google time format" do
-      event = Event.new(
-        start_time: '2013-05-13T05:00:00z',
-        end_time: '2013-05-13T15:30:00z',
-      )
-
-      expect(helper.format_google_timespan(event)).to eq \
-        "20130513T050000Z/20130513T153000Z"
-    end
-
-    it "should the end time to the start time" do
-      event = Event.new(start_time: '2014-07-26T13:00:00-0700')
-      expect(helper.format_google_timespan(event)).to eq \
-        "20140726T200000Z/20140726T200000Z"
     end
   end
